perf: authHeader verify
