+++ apps/backend-mock/api/auth/codes.ts
@@ -1,15 +1,14 @@
 import { verifyAccessToken } from '~/utils/jwt_utils';
 
 export default eventHandler((event) => {
-  const token = getHeader(event, 'Authorization');
-  if (!token) {
+  const userinfo = verifyAccessToken(event);
+  if (!userinfo) {
     setResponseStatus(event, 401);
     return useResponseError('UnauthorizedException', 'Unauthorized Exception');
   }
 
-  const { username } = verifyAccessToken(token);
   const codes =
-    MOCK_CODES.find((item) => item.username === username)?.codes ?? [];
+    MOCK_CODES.find((item) => item.username === userinfo.username)?.codes ?? [];
 
   return useResponseSuccess(codes);
 });
