+++ apps/backend-mock/api/menu/all.ts
@@ -1,16 +1,13 @@
 import { verifyAccessToken } from '~/utils/jwt_utils';
 
 export default eventHandler((event) => {
-  const token = getHeader(event, 'Authorization');
-
-  if (!token) {
+  const userinfo = verifyAccessToken(event);
+  if (!userinfo) {
     setResponseStatus(event, 401);
     return useResponseError('UnauthorizedException', 'Unauthorized Exception');
   }
 
-  const { username } = verifyAccessToken(token);
-
   const menus =
-    MOCK_MENUS.find((item) => item.username === username)?.menus ?? [];
+    MOCK_MENUS.find((item) => item.username === userinfo.username)?.menus ?? [];
   return useResponseSuccess(menus);
 });
