+++ apps/backend-mock/api/user/info.ts
@@ -1,13 +1,7 @@
 import { verifyAccessToken } from '~/utils/jwt_utils';
 
 export default eventHandler((event) => {
-  const token = getHeader(event, 'Authorization');
-  if (!token) {
-    setResponseStatus(event, 401);
-    return useResponseError('UnauthorizedException', 'Unauthorized Exception');
-  }
-
-  const userinfo = verifyAccessToken(token);
+  const userinfo = verifyAccessToken(event);
   if (!userinfo) {
     setResponseStatus(event, 401);
     return useResponseError('UnauthorizedException', 'Unauthorized Exception');
