+++ packages/effects/request/src/request-client/request-client.ts
@@ -99,7 +99,8 @@ class RequestClient {
         const authorization = this.makeAuthorization?.(config);
         if (authorization) {
           const { token } = authorization.tokenHandler?.() ?? {};
-          config.headers[authorization.key || 'Authorization'] = token;
+          config.headers[authorization.key || 'Authorization'] =
+            `Bearer ${token}`;
         }
 
         const requestHeader = this.makeRequestHeaders?.(config);
